% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accdb_to_csv.R
\name{accdb_to_csv}
\alias{accdb_to_csv}
\title{accdb_to_csv}
\usage{
accdb_to_csv(accdb_file, output_dir, ...)
}
\arguments{
\item{accdb_file}{path to access database}

\item{output_dir}{directory to output the results}

\item{...}{arguments to pass to fwrite}
}
\value{
vector of table names
}
\description{
accdb_to_csv
}
